third_party/blink/web_tests/virtual/plz-dedicated-worker-disabled/external/wpt/html/cross-origin-embedder-policy/dedicated-worker-cache-storage.https-expected.txt

This is a testharness.js-based test.
[FAIL] coep-none coep-require-corp corp-undefined
  assert_equals: expected "failure" but got "success"
Harness: the test ran to completion.
